+++ eng/code-quality-reports/src/main/java/com/azure/tools/checkstyle/checks/GoodLoggingCheck.java
@@ -10,12 +10,9 @@ import com.puppycrawl.tools.checkstyle.api.TokenTypes;
 import com.puppycrawl.tools.checkstyle.utils.TokenUtil;
 
 import java.util.ArrayDeque;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.Objects;
 import java.util.Queue;
-import java.util.Set;
 
 /**
  * Good Logging Practice:
@@ -29,9 +26,6 @@ import java.util.Set;
  * </ol>
  */
 public class GoodLoggingCheck extends AbstractCheck {
-    private static final String CLIENT_LOGGER_PATH = "com.azure.core.util.logging.ClientLogger";
-    private static final String CLIENT_LOGGER = "ClientLogger";
-    private static final String LOGGER = "logger";
     private static final int[] REQUIRED_TOKENS = new int[]{
         TokenTypes.IMPORT,
         TokenTypes.INTERFACE_DEF,
@@ -52,9 +46,57 @@ public class GoodLoggingCheck extends AbstractCheck {
     // A LIFO queue stores the class names, pop top element if exist the class name AST node
     private final Queue<String> classNameDeque = Collections.asLifoQueue(new ArrayDeque<>());
     // Collection of Invalid logging packages
-    private static final Set<String> INVALID_LOGS = Collections.unmodifiableSet(new HashSet<>(Arrays.asList(
+    private static final String[] INVALID_LOGS = new String[] {
         "org.slf4j", "org.apache.logging.log4j", "java.util.logging"
-    )));
+    };
+
+    private String fullyQualifiedLoggerName = "com.azure.core.util.logging.ClientLogger";
+    private String simpleClassName = "ClientLogger";
+    private String loggerName = "logger";
+
+    /**
+     * Sets the fully qualified logger name.
+     * <p>
+     * If not set this will default to {@code com.azure.core.util.logging.ClientLogger}.
+     *
+     * @param fullyQualifiedLoggerName the fully qualified logger name.
+     * @throws IllegalArgumentException if the fully qualified logger name is null or empty.
+     */
+    public final void setFullyQualifiedLoggerName(String fullyQualifiedLoggerName) {
+        if (fullyQualifiedLoggerName == null || fullyQualifiedLoggerName.isEmpty()) {
+            throw new IllegalArgumentException("fullyQualifiedLoggerName cannot be null or empty.");
+        }
+
+        this.fullyQualifiedLoggerName = fullyQualifiedLoggerName;
+    }
+
+    /**
+     * Sets the simple class name for the logger.
+     * <p>
+     * If not set this will default to {@code ClientLogger}.
+     *
+     * @param simpleClassName the simple class name for the logger.
+     * @throws IllegalArgumentException if the simple class name is null or empty.
+     */
+    public final void setSimpleClassName(String simpleClassName) {
+        if (simpleClassName == null || simpleClassName.isEmpty()) {
+            throw new IllegalArgumentException("simpleClassName cannot be null or empty.");
+        }
+
+        this.simpleClassName = simpleClassName;
+    }
+
+    /**
+     * Sets the case-insensitive name of the logger instance.
+     * <p>
+     * If not set this will default to {@code logger}.
+     *
+     * @param loggerName the case-insensitive name of the logger instance.
+     * @throws IllegalArgumentException if the logger name is null or empty.
+     */
+    public final void setLoggerName(String loggerName) {
+        this.loggerName = loggerName;
+    }
 
     @Override
     public int[] getDefaultTokens() {
@@ -90,13 +132,13 @@ public class GoodLoggingCheck extends AbstractCheck {
         switch (ast.getType()) {
             case TokenTypes.IMPORT:
                 final String importClassPath = FullIdent.createFullIdentBelow(ast).getText();
-                hasClientLoggerImported = hasClientLoggerImported || importClassPath.equals(CLIENT_LOGGER_PATH);
+                hasClientLoggerImported = hasClientLoggerImported || importClassPath.equals(fullyQualifiedLoggerName);
 
-                INVALID_LOGS.forEach(item -> {
-                    if (importClassPath.startsWith(item)) {
-                        log(ast, String.format(NOT_CLIENT_LOGGER_ERROR, "external logger", CLIENT_LOGGER_PATH, item));
+                for (String invalidLog : INVALID_LOGS) {
+                    if (importClassPath.startsWith(invalidLog)) {
+                        log(ast, String.format(NOT_CLIENT_LOGGER_ERROR, "external logger", fullyQualifiedLoggerName, invalidLog));
                     }
-                });
+                }
                 break;
             case TokenTypes.CLASS_DEF:
             case TokenTypes.INTERFACE_DEF:
@@ -116,7 +158,7 @@ public class GoodLoggingCheck extends AbstractCheck {
                 }
                 final String methodCallName = FullIdent.createFullIdentBelow(dotToken).getText();
                 if (methodCallName.startsWith("System.out") || methodCallName.startsWith("System.err")) {
-                    log(ast, String.format(NOT_CLIENT_LOGGER_ERROR, "Java System", CLIENT_LOGGER_PATH, methodCallName));
+                    log(ast, String.format(NOT_CLIENT_LOGGER_ERROR, "Java System", fullyQualifiedLoggerName, methodCallName));
                 }
                 break;
             default:
@@ -137,7 +179,7 @@ public class GoodLoggingCheck extends AbstractCheck {
             return false;
         }
         return TokenUtil.findFirstTokenByPredicate(typeAST, node ->
-            node.getType() == TokenTypes.IDENT && node.getText().equals(CLIENT_LOGGER)
+            node.getType() == TokenTypes.IDENT && node.getText().equals(simpleClassName)
         ).isPresent();
     }
 
@@ -149,7 +191,7 @@ public class GoodLoggingCheck extends AbstractCheck {
     private void checkLoggerInstantiation(DetailAST literalNewToken) {
         final DetailAST identToken = literalNewToken.findFirstToken(TokenTypes.IDENT);
         // Not ClientLogger instance
-        if (identToken == null || !identToken.getText().equals(CLIENT_LOGGER)) {
+        if (identToken == null || !identToken.getText().equals(simpleClassName)) {
             return;
         }
         // LITERAL_NEW node always has ELIST node below
@@ -180,8 +222,8 @@ public class GoodLoggingCheck extends AbstractCheck {
         }
         // Check if the Logger instance named as 'logger/LOGGER'.
         final DetailAST identAST = varToken.findFirstToken(TokenTypes.IDENT);
-        if (identAST != null && !identAST.getText().equalsIgnoreCase(LOGGER)) {
-            log(varToken, String.format(LOGGER_NAME_ERROR, LOGGER, identAST.getText()));
+        if (identAST != null && !identAST.getText().equalsIgnoreCase(loggerName)) {
+            log(varToken, String.format(LOGGER_NAME_ERROR, loggerName, identAST.getText()));
         }
     }
 }