+++ sdk/clientcore/core/src/main/java/module-info.java
@@ -26,6 +26,9 @@ module io.clientcore.core {
     exports io.clientcore.core.util.configuration;
     exports io.clientcore.core.util.serializer;
     exports io.clientcore.core.util.auth;
+    exports io.clientcore.core.instrumentation;
+    exports io.clientcore.core.instrumentation.tracing;
+    exports io.clientcore.core.instrumentation.logging;
 
     uses io.clientcore.core.http.client.HttpClientProvider;