+++ sdk/resourcemanager/azure-resourcemanager-dns/src/main/java/com/azure/resourcemanager/dns/models/AaaaRecordSet.java
@@ -8,6 +8,10 @@ import java.util.List;
 /** An immutable client-side representation of a AAAA (IPv6) record set in Azure DNS Zone. */
 @Fluent
 public interface AaaaRecordSet extends DnsRecordSet {
-    /** @return the IPv6 addresses of AAAA records in this record set */
+    /**
+     * Gets the IPv6 addresses of AAAA records in this record set.
+     *
+     * @return the IPv6 addresses of AAAA records in this record set
+     */
     List<String> ipv6Addresses();
 }